Initial import
Though not tested, this release should work for a restful model (in which the start_login and complete_login methods are added as collection get to the restful route).
It will forward to the index-method of the controller and will not do anything fancy.

The license is bound to whatever ruby-openid allows you to do... The LICENSE file is adopted accordingly.